Added Melee weapon and inheritance for Shield
