+++ objects/weapon/Shield.py
@@ -4,33 +4,18 @@ import pygame.math
 import pygame.transform
 from config.Config import *
 from objects.weapon.Bullet import Bullet
+from objects.weapon.MeleeWeapon import MeleeWeapon
 
-class Shield(pygame.sprite.Sprite):
+
+class Shield(MeleeWeapon):
     def __init__(self, level, groups, image_path, owner, owner_distance, cooldown):
-        super().__init__(groups)
-        self.level = level
-        self.start_image = pygame.image.load(image_path).convert_alpha()
-        self.image = pygame.image.load(image_path).convert_alpha()
-        self.rect = self.image.get_rect()
-        self.owner = owner
-        self.owner_distance = owner_distance
-        self.last_use = cooldown
-        self.cooldown = cooldown
-        self.reflect = False
+        super().__init__(level, groups, image_path, owner, owner_distance, cooldown)
         self.reflect2 = False
-
-    def move(self):
-        self.image = pygame.transform.rotate(self.start_image, self.owner.look_angle.angle_to(pygame.math.Vector2(0, 1)))
-        if not self.reflect:
-            self.image.set_alpha(SHIELD_ALPHA)  # later add animation
-        self.rect = self.image.get_rect()
-        self.pos = self.owner.pos + self.owner.look_angle * self.owner_distance
-        self.rect.center = self.pos
 
     def reflect_bullets(self):
         if self.last_use >= self.cooldown:
             self.last_use = 0
-            self.reflect = True
+            self.use = True
 
     def redirect_bullet(self, bullet: Bullet):
         new_speed = self.owner.look_angle * bullet.speed.length() * BULLET_REFLECTION_ACCELERATION
@@ -38,11 +23,11 @@ class Shield(pygame.sprite.Sprite):
         bullet.damage *= BULLET_REFLECTION_DAMAGE_UP
 
     def update(self, dt):
-        self.last_use += dt
+        super().update(dt)
         self.move()
         # check order of Level.bullets_update and Level.visible.update - reflect2 is necessary because of this order
         if self.reflect2:
-            self.reflect = False
+            self.use = False
             self.reflect2 = False
-        elif self.reflect:
+        elif self.use:
             self.reflect2 = True
